Update CSS file
The minimum height of the page to correctly display the footer is fixed.
The first element of a page has no top margin but has a top padding
instead.

Added a font.

Set the minimum width the page should be rendered at.

Some of the links in the header disappear if the page is too small.

Removed some old files.